fix: move tests to src/_data/ and restore SRI hash on Umami script
- Verplaats testbestanden van tests/ naar src/_data/ conform CLAUDE.md
- Pas vitest.config.js aan naar src/**/*.test.js
- Herstel SRI integrity hash op cloud.umami.is script (sha384)
