+++ src/_data/helpers.test.js
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import helpers from '../src/_data/helpers.js';
+import helpers from './helpers.js';
 
 describe('helpers', () => {
   describe('getLinkActiveState', () => {
