+++ src/_includes/layouts/base.html
@@ -45,7 +45,7 @@
     <link href="https://bsky.app/profile/groenoranjerood.bsky.social" rel="me">
 
     <script src="/assets/js/main.js" defer></script>
-    <script defer src="https://cloud.umami.is/script.js" data-website-id="6b5a5a34-b2f6-4e32-b181-fab14b35f53e"></script>
+    <script defer src="https://cloud.umami.is/script.js" data-website-id="6b5a5a34-b2f6-4e32-b181-fab14b35f53e" integrity="sha384-yRsxKUe1uwvKWyAzDiU58FsIbO8orQJUzy1kPcDQAfAuaczzL5MxTlI4K/AABHhq" crossorigin="anonymous"></script>
 </head>
 
 <body>
