+++ vitest.config.js
@@ -2,6 +2,6 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
     test: {
-        include: ['tests/**/*.test.js'],
+        include: ['src/**/*.test.js'],
     },
 });
